requirements.txt
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
babel==2.18.0
celery[redis]==5.4.0
celery-redbeat==2.2.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
courlan==1.3.2
cryptography==46.0.7
dateparser==1.4.0
distro==1.9.0
dnspython==2.8.0
greenlet==3.4.0
h11==0.16.0
htmldate==1.9.4
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jiter==0.14.0
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jusText==3.0.2
langchain==1.2.15
langchain-classic==1.0.3
langchain-core==1.2.29
langchain-mcp-adapters==0.2.2
langchain-mongodb==0.11.0
langchain-openai==1.1.13
langchain-text-splitters==1.1.1
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-checkpoint-mongodb==0.3.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.13
langsmith==0.7.31
lark==1.3.1
lxml==6.0.4
lxml_html_clean==0.4.4
mcp==1.27.0
numpy==2.4.4
openai==2.31.0
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.1
pycparser==3.0
pydantic==2.13.0
pydantic-settings==2.13.1
pydantic_core==2.46.0
PyJWT==2.12.1
pymongo==4.15.5
pymongo-search-utils==0.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
rpds-py==0.30.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
sse-starlette==3.3.4
starlette==1.0.0
tenacity==9.1.4
tiktoken==0.12.0
tld==0.13.2
tqdm==4.67.3
trafilatura==2.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.44.0
voluptuous==0.16.0
xxhash==3.6.0
zstandard==0.25.0
